adding all the gulp details
