+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules/
+public/styles/
+public/scripts/
