+++ index.html
@@ -5,6 +5,8 @@
 	<title>Nomad App</title>
 </head>
 <body>
+	<h1>hello</h1>
+	<h2>goodbye</h2>
 	
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>	
 	<script src="script.js"></script>
